chore: use console.log where appropriate. Move log.js
