+++ CICD/compareTestResults.js
@@ -1,6 +1,6 @@
 const fs = require("fs");
 
-const log = require("../src/log.js");
+const log = require("../src//common/log");
 
 log.setLogLevel(log.logLevels.verbose);
